chore: moved preference schema definitions to a different file to avoid compass specific imports in provider
