+++ packages/compass-preferences-model/package.json
@@ -48,12 +48,13 @@
     "reformat": "npm run eslint . -- --fix && npm run prettier -- --write ."
   },
   "dependencies": {
-    "@mongodb-js/compass-user-data": "^0.1.14",
     "@mongodb-js/compass-logging": "^1.2.11",
+    "@mongodb-js/compass-user-data": "^0.1.14",
     "bson": "^6.2.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
-    "yargs-parser": "^21.1.1"
+    "yargs-parser": "^21.1.1",
+    "zod": "^3.22.3"
   },
   "devDependencies": {
     "@mongodb-js/compass-user-data": "^0.1.14",
